% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_gsod_web.R
\name{handle_gsod_web}
\alias{handle_gsod_web}
\title{Download climate data from the Global Summary of the Day database (faster)}
\usage{
handle_gsod_web(
  action,
  location = NULL,
  time_interval = NULL,
  path = "climate_data",
  end_at_present = FALSE,
  stations_to_choose_from = 25,
  max_distance = 150,
  min_overlap = 0,
  drop_most = TRUE,
  update_all = FALSE,
  update_station_list = FALSE,
  add.Date = TRUE,
  clean_up = NULL,
  verbose = "normal"
)
}
\arguments{
\item{action}{Can be set to:
\enumerate{
\item "\code{list_stations}" to download a list of stations in proximity of a
location.
\item "\code{download_weather}" to download climate data for a specific time
interval from one or more stations listed in the downloaded stations
list.
\item "\code{delete}" to clean up the path used for downloading. Can be used with
\code{clean_up = "all"} to delete the entire path, or with
\code{clean_up = "station"} in combination with the parameter \code{location} set
to the station \code{chillR_code} to be deleted, as used in
\code{action = "download_weather"}.
\item \code{raw_data_set}to convert the data into \code{chillR} format. This converts
the units into SI units. For the raw data documentation visit
https://www.ncei.noaa.gov/data/global-summary-of-the-day/doc/readme.txt
}}

\item{location}{Needs to be set to either a geographic location
(\code{c(longitude,latitude)}) in conjunction with \code{action = "list_stations"}
or \code{chillR_code(s)} extracted from the station list.}

\item{time_interval}{Normally a numeric vector with two elements, specifying
the start and end date of the period of interest. Only required when
running \code{action = list_stations} or \code{action = download_weather}.
Can also be a numeric vector with only one element, if used with
\code{end_at_present = TRUE}.}

\item{path}{Default value is "climate_data". Here the path for the downloads
can be set manually if desired. The function will then concatenate this
string with the current working directory path. Use \code{getwd()} and
\code{setwd()} if you are not sure what your current working directory is.}

\item{end_at_present}{Default is \code{FALSE}. Can be set to \code{TRUE}, if only one
element is provided in \code{time_interval}. In this case the second value will
be exacted from the system time.}

\item{stations_to_choose_from}{Default is 25. This is limiting the amount of
stations outputted while using \code{action = list_stations}.The
\code{max_distance} parameter might be too low, if the output does contain less
then 25 stations.}

\item{max_distance}{Default is 150 kilometer and freely adjustable. This is
limiting the amount of stations outputted while using
\code{action = list_stations}.}

\item{min_overlap}{Default is 0. This is used to create the \code{Overlap_years}
and \code{Perc_interval_covered} columns in the station list file. Can be set
higher to filter for stations that cover a certain amount of the desired
\code{time_interval}. Setting this too high might result in a lot of available
stations to be lost.}

\item{drop_most}{Default is \code{TRUE}. This is used in combination with
\code{action = raw_data_set} to convert the raw downloaded data into \code{chillR}
format. Can be set to \code{FALSE} to keep the overall structure of the raw
data and only convert into SI units.}

\item{update_all}{Default is \code{FALSE}. This is used to speed up the function
run time by only downloading data that is missing in the \code{path}
directories. Can be set to \code{TRUE} if all the data is supposed to be
re-downloaded.}

\item{update_station_list}{Default is \code{FALSE}. Similar behavior as
\code{update_all}. The station list changes much less than the downloaded data,
but it is still advised to update it from time to time.}

\item{add.Date}{Default is \code{TRUE}. This is used in combination with
\code{action = raw_data_set} and will add missing dates to the \code{chillR}
formated data set. These dates will have \code{NAs} in the data columns.}

\item{clean_up}{This is used in Combination with \code{action = delete}. Can be
either \code{clean_up = "all"} to delete the entire download path set in
\code{path}, or \code{clean_up = "station"} to delete specific station folders.
\code{clean_up = "station"} needs to be used in combination with the parameter
\code{location} set to the station(s) \code{chillR_codes}, as used in
\code{action = "download_weather"}.}

\item{verbose}{Default is "\code{normal}" and can also be set to "\code{detailed}" or
"\code{quiet}". Set to "\code{normal}" it will still give download progress updates, but
keep the console comments to a minimum. setting it to "\code{detailed}" can
helpful for debugging problems. "\code{quiet}" will give no console updates.}
}
\description{
This function's main features are downloading and converting climate data from
the Global Summary of the Day database (GSOD) from the National Climatic
Data Centre (NCDC) of the National Oceanic and Atmospheric Administration
(NOAA).

The standard workflow for this is:
\enumerate{
\item Getting the list of stations that are close to a specified position
using the \code{action = "list_stations"} in combination with the
\code{location = c(longitude,latitude)} parameter. The parameter
\code{time_interval = c(start_year,end_year)} can be used to see the
coverage of each station.
\item Downloading raw climate data from suitable stations by setting the
parameter \code{action = "download_weather"} and selecting the stations
\code{chillR_code} from the stationlist with
\code{location = stationlist$chillR_code[n]}. In addition the parameter
\code{time_interval} needs to be set to a desired timeinterval as in the
previous step.
\item Converting the raw data into a suitable \code{chillR} format by setting
the \code{action} parameter equal to the raw data set.(\code{action = raw_data})
}

There are many other parameters that can be set - see the Arguments
section for more information.
}
\details{
This function is an alternative to the original \code{handle_gsod()} function from
the \code{chillR} package and strives to run faster and have some additional
tweaks. The increased speed stems mainly from saving the data to the disk
and also loading it from there if already present. It will only
re-download ALL the data if the parameter \code{update_all} is set to \code{TRUE}.
One major difference between this function and the original is that
the downloaded data is initially kept in its raw form (as found in the
GSOD database). This allows for own manipulation of the data, but keep in
mind that the units are not yet converted to SI. Setting the \code{action}
parameter to the raw data set and also setting \code{drop_most} to \code{FALSE}
will keep the raw data structure but convert everything into SI units.
Setting \code{drop_most} to \code{TRUE} will result in the \code{chillR} format.

The units are converted as following:
\itemize{
\item TEMP   - Mean temperature             (.1 Fahrenheit into Celsius)
\item DEWP   - Mean dew point               (.1 Fahrenheit into Celsius)
\item SLP    - Mean sea level pressure      (.1 mb kept as millibar or hektoPascal)
\item STP    - Mean station pressure        (.1 mb kept as millibar or hektoPascal)
\item VISIB  - Mean visibility              (.1 miles into kilometer)
\item WDSP   - Mean wind speed              (.1 knots into meter per second)
\item MXSPD  - Maximum sustained wind speed (.1 knots into meter per second)
\item GUST   - Maximum wind gust            (.1 knots into meter per second)
\item MAX    - Maximum temperature          (.1 Fahrenheit into Celsius)
\item MIN    - Minimum temperature          (.1 Fahrenheit into Celsius)
\item PRCP   - Precipitation amount         (.01 inches into millimeter)
\item SNDP   - Snow depth                   (.1 inches into millimeter)
\item FRSHTT - Indicator for occurrence of:
\itemize{
\item Fog, Rain or Drizzle, Snow or Ice Pellets
\item Hail, Thunder, Tornado/Funnel Cloud
}
}
}
\note{
The GSOD database is described here: \url{https://www.ncei.noaa.gov/access/metadata/landing-page/bin/iso?id=gov.noaa.ncdc:C00516#Documentation}

Many databases have data quality flags, which may sometimes indicate that
data isn't reliable. These are not considered by this function, but the
relevant columns can still be found in the \verb{raw data}. Quality control
columns include "Attribute", please refer to the GSOD documentation above
for a more detailed explanation.
}
\examples{
\dontrun{

## stationlist download
# Bonn
lat <- 50.7341602
long <- 7.0871843
stationlist <-
  handle_gsod_web(action = "list_stations",
                  time_interval = c(1995,2000),
                  location = c(long,lat) # ,
                  # path = "climate_data",              # standard option, can be set to something different
                  # update_station_list = F,            # standard option, tries to load from disk, if not possible it downloads - can be set to TRUE so it always downloads
                  # max_distance = 150,                 # standard option, can be adjusted
                  # min_overlap = 0,                    # standard option, percentage from 0 - 100, might obscure relevant locations if to high
                  # stations_to_choose_from = 25        # standard option, can be adjusted
                  )


## data download
test_data <-
  handle_gsod_web(action = "download_weather",
                  time_interval = c(1995,2000),
                  location = stationlist$chillR_code[c(1,2)]#,
                  # end_at_present = F,                 # standard option, can be set to true if time_interval has only one value
                  # path = "climate_data",              # standard option, can be set to something different
                  # verbose = "normal",                 # standard option, minimal information
                  # verbose = "detailed",               #                  detailed information
                  # verbose = "quiet",                  #                  no information - does not need to be 'quiet', but different from 'normal' or 'detailed'
                  # update_all = F,                     # standard option, can be set to TRUE, if data should be downloaded again (updated)
                  )


## chillR formating
test_data_clean <- handle_gsod_web(test_data,
                                   # drop_most = T,     # standard option, can be set to false to keep all columns and only convert to SI units
                                   # add.Date = T       # standard option, can be set to FALSE, but not advised if you want datasets without gaps
                                   )


## data deletion on disk
# functions will ask for confirmation in th console - 'y' for yes to confirm deletion, anything else cancels the deletion
handle_gsod_web(action = "delete",
                clean_up = "all")

handle_gsod_web(action = "delete",
                clean_up = "station",
                location = stationlist$chillR_code[c(1,2)])
}

}
\author{
Adrian Fuelle, \email{afuelle@outlook.de};
Eike Luedeling, \email{eike@eikeluedeling.com}
}
